requirements.txt
# Data Collection
google-play-scraper==1.2.4
pandas==2.0.3
numpy==1.26.4

# NLP & ML
scikit-learn==1.3.0
nltk==3.8.1
Sastrawi==1.0.1

# MLOps
dvc==3.20.0
mlflow==2.7.1

# Database
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
pymongo==4.5.0

# Scheduling
schedule==1.2.0
apscheduler==3.10.4

# API & Monitoring
fastapi==0.103.1
uvicorn==0.23.2
prometheus-client==0.17.1

# Visualization
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2
streamlit==1.28.0

# Utils
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.3.0
